ECL-496: Converted revenue to BigDecimal

+++ app/uk/gov/hmrc/economiccrimelevyregistration/connectors/EclCalculatorConnector.scala
@@ -31,7 +31,7 @@ class EclCalculatorConnector @Inject() (appConfig: AppConfig, httpClient: HttpCl
   private val eclCalculatorUrl: String =
     s"${appConfig.eclCalculatorBaseUrl}/economic-crime-levy-calculator"
 
-  def calculateLiability(relevantApLength: Int, relevantApRevenue: Double)(implicit
+  def calculateLiability(relevantApLength: Int, relevantApRevenue: BigDecimal)(implicit
     hc: HeaderCarrier
   ): Future[CalculatedLiability] =
     httpClient.POST[CalculateLiabilityRequest, CalculatedLiability](
